changed the log functions parameters from uint64 to int

+++ zeabus_elec_ros/src/peripheral_bridge.cpp
@@ -149,9 +149,9 @@ static bool b_service_get_depth(    zeabus_utility::ServiceDepth::Request &x_req
 
     // print and publish the log
     v_log_action(   x_publisher_action_log,
-                    ( int64_t )ki_ACTION_GET_DEPTH_CALLED,
-                    ( int64_t )0U,
-                    ( int64_t )0U,
+                    ki_ACTION_GET_DEPTH_CALLED,
+                    ( int )0U,
+                    ( int )0U,
                     x_description );
 
     // response with depth_state
@@ -163,9 +163,9 @@ static bool b_service_get_depth(    zeabus_utility::ServiceDepth::Request &x_req
 
     // print and publish the log
     v_log_action(   x_publisher_action_log,
-                    ( int64_t )ki_ACTION_GET_DEPTH_COMPLETE,
-                    ( int64_t )( x_depth_state.depth * 100U ),
-                    ( int64_t )0U,
+                    ki_ACTION_GET_DEPTH_COMPLETE,
+                    ( int )( x_depth_state.depth * 100U ),
+                    ( int )0U,
                     x_description );
 
     return true;
@@ -190,9 +190,9 @@ static bool b_set_io_pin_state( zeabus_elec_ros::ServiceIOPinState::Request &x_r
 
     // print and publish the log
     v_log_action(   x_publisher_action_log,
-                    ( int64_t )ki_ACTION_SET_IO_PIN_STATE_CALLED,
-                    ( int64_t )x_request.u_io_pin_index,
-                    ( int64_t )x_request.is_io_pin_state_high,
+                    ki_ACTION_SET_IO_PIN_STATE_CALLED,
+                    ( int )x_request.u_io_pin_index,
+                    ( int )x_request.is_io_pin_state_high,
                     x_description );
 
     try
@@ -256,9 +256,9 @@ static bool b_set_io_pin_state( zeabus_elec_ros::ServiceIOPinState::Request &x_r
 
         // print and publish the log
         v_log_action(   x_publisher_action_log,
-                        ( int64_t )ki_ACTION_SET_IO_PIN_STATE_COMPLETE,
-                        ( int64_t )x_request.u_io_pin_index,
-                        ( int64_t )x_request.is_io_pin_state_high,
+                        ki_ACTION_SET_IO_PIN_STATE_COMPLETE,
+                        ( int )x_request.u_io_pin_index,
+                        ( int )x_request.is_io_pin_state_high,
                         x_description );
     }
     catch( const int &ki_error )
@@ -294,8 +294,8 @@ static bool b_set_io_pin_state( zeabus_elec_ros::ServiceIOPinState::Request &x_r
 
         // print and publish the log
         v_log_hardware_error(   x_publisher_hardware_error_log,
-                                ( int64_t ) ki_error,
-                                ( int64_t ) i_hardware_error_code,
+                                ki_error,
+                                i_hardware_error_code,
                                 x_description );
     }
 
@@ -440,8 +440,8 @@ int main( int argc, char** argv )
 
                 // print and publish the log
                 v_log_hardware_error(   x_publisher_hardware_error_log,
-                                        ( int64_t ) ki_error,
-                                        ( int64_t ) i_hardware_error_code,
+                                        ki_error,
+                                        i_hardware_error_code,
                                         x_description );
             }
 
@@ -486,8 +486,8 @@ int main( int argc, char** argv )
 
         // print and publish the log
         v_log_hardware_error_fatal( x_publisher_hardware_error_log,
-                                    ( int64_t )ki_error,
-                                    ( int64_t )i_hardware_error_code,
+                                    ki_error,
+                                    i_hardware_error_code,
                                     x_description );
     }
